Now using input system for Gatherer
