+++ Assets/Scripts/PlayerMovement/WardenMovement.cs
@@ -1,7 +1,4 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
-using static PlayerMovement;
 
 public class WardenMovement : MonoBehaviour
 {
